Delete base64.cpp
This code is included in my running red team agent on my local machine in a project and allows for base64 encoding. I call it in my main red team agent where necessary, but it is not my code and is available on GitHub, I would like to use this repo to display my code only. If you require base64 functionality in C++, please search GitHub, there are many available projects.